% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine_fun_cat.R
\name{wine_fun_cat}
\alias{wine_fun_cat}
\title{Plot function}
\usage{
wine_fun_cat(type = "linear", growth = 1, cat = TRUE)
}
\arguments{
\item{type}{Linear or exponential? Default is type = 'linear' (options are type='linear' or type='exponential')}

\item{growth}{How much does wine effect happiness? Default is growth = 1}

\item{cat}{Do you want a cat background for your plot? Default is TRUE (options are TRUE or FALSE)}
}
\description{
This function generates a plot with linear or exponential growth, with the option of a cat backgroud. The cats package must be installed from https://github.com/hilaryparker/cats (install_github('cats','hilaryparker')) as well as the httr package.
}
\examples{
wine_fun(type='exponential', growth='5')
wine_fun_cat()
}
\keyword{wine}
